==> notebooks/02_Fixing_Datatypes.ipynb <==
import numpy as np
import pandas as pd

input_path = './../data/processed/After_01_.csv'
df = pd.read_csv(input_path)

#### 1. PPI

    - This column has some text values, and doesn't have proper numeric values in the column.
    - Since i was not able to find the right values, i'm dropping these rows.
    - After removing the unnecessary rows, convert it into integer column.


df = df[df['PPI'].str.isnumeric() == True]
df['PPI'] = df['PPI'].astype('int')

output_path = './../data/processed/After_02_Fixing_Datatypes.csv'
df.to_csv(output_path, index = False)

df.columns
