==> notebooks/06_Filling_Missing_Values.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

import sys
import os

# Add the root of your project to sys.path
project_root = os.path.abspath('..')  # or '.' if your notebook is in root
if project_root not in sys.path:
    sys.path.insert(0, project_root)

%load_ext autoreload
%autoreload 2

import importlib
import src.eda
importlib.reload(src.eda)

input_path = './../data/processed/After_05_Fixing_Categories.csv'
df = pd.read_csv(input_path)

### Numerical Columns


Numeric(Missing values) : thickness, Battery_Capacity, CPU_Transistor_Size,

Numeric(transformation) : thickness, Battery_Capacity


df.isna().sum().sort_values(ascending = False).head(10)

X = df.drop(columns = 'Price_In_Dollars')
y = df['Price_In_Dollars']

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.1)

#### Filling Missing values


import numpy as np

def fillMissingValuesWhole(X_train, X_test, columnToUse, columnToFill, isNumericColumn):
    mapper = {}

    for brand in X_train[columnToUse].unique():
        if not isNumericColumn:
            fillingValue = X_train[X_train[columnToUse] == brand][columnToFill].mode()
            if fillingValue.empty:
                fillingValue = X_train[columnToFill].mode()
            mapper[brand] = fillingValue[0]
        else:
            fillingValue = X_train[X_train[columnToUse] == brand][columnToFill].mean()
            if np.isnan(fillingValue):
                fillingValue = X_train[columnToFill].mean()
            mapper[brand] = fillingValue

    def transformationFunction(row):
        row[columnToFill] = mapper.get(row[columnToUse], np.nan)
        return row

    def findMissingValuesAndFill(df, column):
        missing_rows = df[df[column].isna()].index
        df.loc[missing_rows, column] = df.loc[missing_rows].apply(transformationFunction, axis=1)[column]

    findMissingValuesAndFill(X_train, columnToFill)
    findMissingValuesAndFill(X_test, columnToFill)

fillMissingValuesWhole(X_train, X_test, 'Brand', 'camera_panorama', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'GPU_company', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'CPU_Model', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'CPU_Brand', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'has_nfc', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'Number_of_cores', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'USB_Type', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'is_foldable_phone', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'Is_OS_Upgradable', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'Battery_Type_Lithium', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'camera_hdr', False)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'number_of_2g_bands', False)

fillMissingValuesWhole(X_train, X_test, 'Brand', 'thickness', True)
fillMissingValuesWhole(X_train, X_test, 'Brand', 'Battery_Capacity', True)

X_train.head()

X_train = X_train.drop(columns = ['Name', 'Model', 'CPU_Model'])
X_test = X_test.drop(columns = ['Name', 'Model', 'CPU_Model'])

X_train.to_csv('./../data/processed/X_train.csv', index = False)
X_test.to_csv('./../data/processed/X_test.csv', index = False)
y_train.to_csv('./../data/processed/y_train.csv', index = False)
y_test.to_csv('./../data/processed/y_test.csv', index = False)
