==> notebooks/07_Feature_Selection.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import OrdinalEncoder, OneHotEncoder
from sklearn.impute import KNNImputer, SimpleImputer
from category_encoders import TargetEncoder
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor
from scipy.stats import boxcox
from sklearn.preprocessing import PowerTransformer
from sklearn.compose import TransformedTargetRegressor

import sys
import os

# Add the root of your project to sys.path
project_root = os.path.abspath('..')  # or '.' if your notebook is in root
if project_root not in sys.path:
    sys.path.insert(0, project_root)

%load_ext autoreload
%autoreload 2

import importlib
import src.eda
importlib.reload(src.eda)

### Read the input


input_path = './../data/processed/After_05_Fixing_Categories.csv'
X = pd.read_csv('./../data/processed/X_train.csv')
y = pd.read_csv('./../data/processed/y_train.csv').squeeze()

X_test = pd.read_csv('./../data/processed/X_test.csv')
y_test = pd.read_csv('./../data/processed/y_test.csv').squeeze()

### Apply Transformations Using a Pipeline


from src.utilities import CustomColumnTransformer

ordinalEncodingColumns = {
    'Sound_3.5mmjack' : ['Yes', 'No'],
    'is_foldable_phone' : [False, True],
    'has_memory_card' : [False, True],
    'memory_type_microSDXC' : [False, True],
    'memory_type_microSDHC' : [False, True],
    'Battery_Is_removable' : ['Removable', 'Non-Removable'],
    'Battery_Type_Lithium' : ['Lithium Ion', 'Lithium Polymer'],
    'Sound_has_stereo_speakers' : ['Yes', 'No'],
    'Number_of_cores' : ['Low_Core', 'Mid_Core', 'High_Core'],
    'CPU_Transistor_Size' : ['> 20nm', '10nm - 20nm', '5nm - 10nm', 'Less than 5nm'],
    'Is_OS_Upgradable' : [False, True],
    'Resolution_Width' : ['Low', 'Medium', 'High'],
    'Resolution_Height' : ['Low', 'Medium', 'High'],
    'max_resolution' : ['720p', '1080p', '1440p', '4k', '8k'],
    'ram' : ['<= 4', '6 - 8', '8-16', '16+'],
    'maincamera_mp' : ['Low', 'Medium', 'High', 'Ultra'],
    'selfiecamera_mp' : ['Low', 'Medium', 'High', 'Ultra'],
    'Bluetooth_version' : ['Less than 4', '5', '6'],
}

targetEncodingColumns = ['Brand', 'GPU_company', 'CPU_Brand', 'USB_Type', 'operating_system']

transformingColumns = ['Total_Pixels', 'Battery_Capacity', 'PPI',	'Display_Size']

pipeline= Pipeline([ ('preprocessing', CustomColumnTransformer(ordinalEncodingColumns, targetEncodingColumns, transformingColumns))])

X['remainder__latest_wifi_version'].value_counts()

X = pd.DataFrame(pipeline.fit_transform(X, y), columns = pipeline.get_feature_names_out())

### Technique 1 : Correlation analysis


temp = X.copy()
temp['Price'] = y

fs1 = temp.corr()['Price'].sort_values(ascending = False).reset_index().rename(columns = {'index' : 'feature', 'Price' : 'correlation'})

### Technique 2 : Random Forest Feature Selection


from sklearn.ensemble import RandomForestRegressor

rf = RandomForestRegressor()

rf.fit(X, y)

fs2 = pd.DataFrame({'feature' : X.columns ,'RF' : rf.feature_importances_})

### Technique 3 : Gradient Boosting Feature Importance


from sklearn.ensemble import GradientBoostingRegressor

gb = GradientBoostingRegressor()

gb.fit(X, y)

fs3 = pd.DataFrame({'feature' : X.columns ,'GB' : gb.feature_importances_})

### Technique 4 : Permutation Importance


from sklearn.inspection import permutation_importance
from sklearn.model_selection import train_test_split

X_train_label, X_test_label, y_train_label, y_test_label = train_test_split(X, y, test_size=0.2, random_state=42)

# Train a Random Forest regressor on label encoded data
rf_label = RandomForestRegressor(n_estimators=100, random_state=42)
rf_label.fit(X_train_label, y_train_label)

# Calculate Permutation Importance
perm_importance = permutation_importance(rf_label, X_test_label, y_test_label, n_repeats=30, random_state=42)

# Organize results into a DataFrame
fs4 = pd.DataFrame({
    'feature': X.columns,
    'permutation_importance': perm_importance.importances_mean
}).sort_values(by='permutation_importance', ascending=False)



fs4

### Technique 5 : RFE


from sklearn.feature_selection import RFE

# Initialize the base estimator
estimator = RandomForestRegressor()

# Apply RFE on the label-encoded and standardized training data
selector_label = RFE(estimator, n_features_to_select=X.shape[1], step=1)
selector_label = selector_label.fit(X, y)

# Get the selected features based on RFE
selected_features = X.columns[selector_label.support_]

# Extract the coefficients for the selected features from the underlying linear regression model
selected_coefficients = selector_label.estimator_.feature_importances_

# Organize the results into a DataFrame
fs5 = pd.DataFrame({
    'feature': selected_features,
    'rfe_score': selected_coefficients
}).sort_values(by='rfe_score', ascending=False)

### Merge all the dataframes with different feature importances


final_fi_df = fs1.merge(fs2,on='feature').merge(fs3,on='feature').merge(fs4,on='feature').merge(fs5,on='feature').set_index('feature')

### Sort the features based on their importance


final_fi_df = (final_fi_df.sum(axis = 1) / final_fi_df.shape[1]).sort_values(ascending = False).reset_index()

### Changing the column names


final_fi_df.columns = ['feature', 'importance']

final_fi_df

#### Selecting all the features which has a minimum feature importance


selected_features = final_fi_df[final_fi_df['importance'] > -100]['feature'].tolist()

selected_features

X = X[selected_features]

X.shape

X.head().iloc[:, 5:]

X.columns
